Support Mathics3 10.0.0
