+++ pkgs/sagemath-fricas/known-test-failures.json
@@ -1705,7 +1705,7 @@
     "sage.interfaces.mathematica": {
         "ntests": 25
     },
-    "sage.interfaces.mathics": {
+    "sage.interfaces.mathics3": {
         "ntests": 148
     },
     "sage.interfaces.maxima": {
