+++ pkgs/sagemath-maxima/known-test-failures.json
@@ -1514,7 +1514,7 @@
     "sage.interfaces.mathematica": {
         "ntests": 25
     },
-    "sage.interfaces.mathics": {
+    "sage.interfaces.mathics3": {
         "ntests": 30
     },
     "sage.interfaces.maxima": {
