+++ pkgs/sagemath-standard/known-test-failures.json
@@ -3692,7 +3692,7 @@
     "sage.interfaces.mathematica": {
         "ntests": 25
     },
-    "sage.interfaces.mathics": {
+    "sage.interfaces.mathics3": {
         "ntests": 30
     },
     "sage.interfaces.matlab": {
