+++ pkgs/sagemath-symbolics/MANIFEST.in
@@ -34,7 +34,7 @@ include sage/matrix/matrix_symbolic_*.p*
 include sage/interfaces/magma*.p*
 include sage/interfaces/maple*.p*
 include sage/interfaces/mathematica.p*
-include sage/interfaces/mathics.p*
+include sage/interfaces/mathics3.p*
 include sage/interfaces/sympy*.p*
 include sage/interfaces/tides.p*
