+++ pkgs/sagemath-symbolics/known-test-failures--standard.json
@@ -1560,7 +1560,7 @@
     "sage.interfaces.mathematica": {
         "ntests": 25
     },
-    "sage.interfaces.mathics": {
+    "sage.interfaces.mathics3": {
         "ntests": 30
     },
     "sage.interfaces.maxima": {
