+++ src/doc/en/reference/interfaces/index.rst
@@ -132,7 +132,7 @@ and testing to make sure nothing funny is going on).
    sage/interfaces/magma_free
    sage/interfaces/maple
    sage/interfaces/mathematica
-   sage/interfaces/mathics
+   sage/interfaces/mathics3
    sage/interfaces/matlab
    sage/interfaces/maxima
    sage/interfaces/maxima_abstract
